chore: move AppliedState to crate raft-store, meta-service request/reply types to crate meta-service
