+++ src/meta/client/Cargo.toml
@@ -23,7 +23,7 @@ common-meta-types = { path = "../types" }
 common-metrics = { path = "../../common/metrics" }
 common-tracing = { path = "../../common/tracing" }
 
-derive_more = "0.99.17"
+derive_more = { workspace = true }
 futures = "0.3.24"
 once_cell = "1.15.0"
 parking_lot = "0.12.1"
