+++ src/meta/raft-store/src/state_machine/client_last_resp.rs
@@ -12,9 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common_meta_types::AppliedState;
 use serde::Deserialize;
 use serde::Serialize;
+
+use crate::applied_state::AppliedState;
 
 /// Client last response that is stored in SledTree
 /// raft state: A mapping of client serial IDs to their state info:
