+++ src/meta/raft-store/src/state_machine/sm.rs
@@ -35,7 +35,6 @@ use common_meta_types::protobuf as pb;
 use common_meta_types::txn_condition;
 use common_meta_types::txn_op;
 use common_meta_types::txn_op_response;
-use common_meta_types::AppliedState;
 use common_meta_types::Change;
 use common_meta_types::Cmd;
 use common_meta_types::ConditionResult;
@@ -71,6 +70,7 @@ use tracing::debug;
 use tracing::error;
 use tracing::info;
 
+use crate::applied_state::AppliedState;
 use crate::config::RaftConfig;
 use crate::key_spaces::ClientLastResps;
 use crate::key_spaces::Expire;
