+++ src/meta/raft-store/tests/it/state_machine/mod.rs
@@ -17,9 +17,9 @@ use std::time::UNIX_EPOCH;
 
 use common_base::base::tokio;
 use common_meta_kvapi::kvapi::KVApi;
+use common_meta_raft_store::applied_state::AppliedState;
 use common_meta_raft_store::state_machine::StateMachine;
 use common_meta_sled_store::openraft;
-use common_meta_types::AppliedState;
 use common_meta_types::Change;
 use common_meta_types::Cmd;
 use common_meta_types::Endpoint;
