+++ src/meta/service/Cargo.toml
@@ -48,6 +48,7 @@ anyhow = { workspace = true }
 async-trait = "0.1.57"
 backon = "0.2.0"
 clap = { workspace = true }
+derive_more = { workspace = true }
 futures = "0.3.24"
 itertools = "0.10.5"
 metrics = "0.20.1"
