+++ src/meta/service/src/meta_service/meta_leader.rs
@@ -16,13 +16,11 @@ use std::sync::Arc;
 
 use common_base::base::tokio::sync::RwLockReadGuard;
 use common_meta_kvapi::kvapi::KVApi;
+use common_meta_raft_store::applied_state::AppliedState;
 use common_meta_raft_store::state_machine::StateMachine;
 use common_meta_sled_store::openraft::error::RemoveLearnerError;
 use common_meta_stoerr::MetaStorageError;
-use common_meta_types::AppliedState;
 use common_meta_types::Cmd;
-use common_meta_types::ForwardRequest;
-use common_meta_types::ForwardResponse;
 use common_meta_types::LogEntry;
 use common_meta_types::MetaDataError;
 use common_meta_types::MetaDataReadError;
@@ -37,10 +35,12 @@ use tracing::debug;
 use tracing::error;
 use tracing::info;
 
+use crate::message::ForwardRequest;
+use crate::message::ForwardRequestBody;
+use crate::message::ForwardResponse;
+use crate::message::JoinRequest;
+use crate::message::LeaveRequest;
 use crate::meta_service::raftmeta::MetaRaft;
-use crate::meta_service::ForwardRequestBody;
-use crate::meta_service::JoinRequest;
-use crate::meta_service::LeaveRequest;
 use crate::meta_service::MetaNode;
 use crate::metrics::ProposalPending;
 use crate::store::RaftStore;
