+++ src/meta/service/src/meta_service/raftmeta.rs
@@ -28,6 +28,7 @@ use common_base::base::tokio::task::JoinHandle;
 use common_base::base::tokio::time::Instant;
 use common_grpc::ConnectionFactory;
 use common_grpc::DNSResolver;
+use common_meta_raft_store::applied_state::AppliedState;
 use common_meta_raft_store::config::RaftConfig;
 use common_meta_raft_store::key_spaces::GenericKV;
 use common_meta_sled_store::openraft;
@@ -39,16 +40,12 @@ use common_meta_stoerr::MetaStorageError;
 use common_meta_types::protobuf::raft_service_client::RaftServiceClient;
 use common_meta_types::protobuf::raft_service_server::RaftServiceServer;
 use common_meta_types::protobuf::WatchRequest;
-use common_meta_types::AppliedState;
 use common_meta_types::Cmd;
 use common_meta_types::ConnectionError;
 use common_meta_types::Endpoint;
 use common_meta_types::ForwardRPCError;
-use common_meta_types::ForwardRequest;
-use common_meta_types::ForwardResponse;
 use common_meta_types::ForwardToLeader;
 use common_meta_types::InvalidReply;
-use common_meta_types::LeaveRequest;
 use common_meta_types::LogEntry;
 use common_meta_types::MetaAPIError;
 use common_meta_types::MetaDataError;
@@ -74,9 +71,12 @@ use tracing::warn;
 use tracing::Instrument;
 
 use crate::configs::Config as MetaConfig;
+use crate::message::ForwardRequest;
+use crate::message::ForwardRequestBody;
+use crate::message::ForwardResponse;
+use crate::message::JoinRequest;
+use crate::message::LeaveRequest;
 use crate::meta_service::meta_leader::MetaLeader;
-use crate::meta_service::ForwardRequestBody;
-use crate::meta_service::JoinRequest;
 use crate::meta_service::RaftServiceImpl;
 use crate::metrics::server_metrics;
 use crate::network::Network;
