+++ src/meta/service/src/store/store_bare.rs
@@ -21,6 +21,7 @@ use anyerror::AnyError;
 use common_base::base::tokio::sync::RwLock;
 use common_base::base::tokio::sync::RwLockWriteGuard;
 use common_exception::WithContext;
+use common_meta_raft_store::applied_state::AppliedState;
 use common_meta_raft_store::config::RaftConfig;
 use common_meta_raft_store::log::RaftLog;
 use common_meta_raft_store::state::RaftState;
@@ -36,7 +37,6 @@ use common_meta_sled_store::openraft::ErrorVerb;
 use common_meta_sled_store::openraft::Membership;
 use common_meta_sled_store::openraft::StateMachineChanges;
 use common_meta_stoerr::MetaStorageError;
-use common_meta_types::AppliedState;
 use common_meta_types::Endpoint;
 use common_meta_types::LogEntry;
 use common_meta_types::MetaError;
