+++ src/meta/service/tests/it/meta_node/meta_node_lifecycle.rs
@@ -28,10 +28,10 @@ use common_meta_types::NodeId;
 use common_meta_types::UpsertKV;
 use databend_meta::configs;
 use databend_meta::init_meta_ut;
-use databend_meta::meta_service::ForwardRequest;
-use databend_meta::meta_service::ForwardRequestBody;
-use databend_meta::meta_service::JoinRequest;
-use databend_meta::meta_service::LeaveRequest;
+use databend_meta::message::ForwardRequest;
+use databend_meta::message::ForwardRequestBody;
+use databend_meta::message::JoinRequest;
+use databend_meta::message::LeaveRequest;
 use databend_meta::meta_service::MetaNode;
 use maplit::btreeset;
 use pretty_assertions::assert_eq;
