+++ src/meta/service/tests/it/tests/service.rs
@@ -26,10 +26,10 @@ use common_meta_client::MetaGrpcClient;
 use common_meta_kvapi::kvapi;
 use common_meta_sled_store::openraft::NodeId;
 use common_meta_types::protobuf::raft_service_client::RaftServiceClient;
-use common_meta_types::ForwardRequest;
-use common_meta_types::ForwardRequestBody;
 use databend_meta::api::GrpcServer;
 use databend_meta::configs;
+use databend_meta::message::ForwardRequest;
+use databend_meta::message::ForwardRequestBody;
 use databend_meta::meta_service::MetaNode;
 use tracing::info;
 use tracing::warn;
