+++ src/meta/types/src/message.rs
@@ -16,130 +16,14 @@ use openraft::raft::AppendEntriesRequest;
 use openraft::raft::InstallSnapshotRequest;
 use openraft::raft::VoteRequest;
 use serde::de::DeserializeOwned;
-use serde::Deserialize;
 use serde::Serialize;
-use thiserror::Error;
 
 use crate::protobuf::RaftReply;
 use crate::protobuf::RaftRequest;
-use crate::AppliedState;
-use crate::Endpoint;
-use crate::GetKVReply;
-use crate::GetKVReq;
 use crate::InvalidReply;
-use crate::ListKVReply;
-use crate::ListKVReq;
 use crate::LogEntry;
-use crate::MGetKVReply;
-use crate::MGetKVReq;
-use crate::NodeId;
 use crate::TxnOpResponse;
 use crate::TxnReply;
-
-#[derive(Error, Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub enum RetryableError {
-    /// Trying to write to a non-leader returns the latest leader the raft node knows,
-    /// to indicate the client to retry.
-    #[error("request must be forwarded to leader: {leader}")]
-    ForwardToLeader { leader: NodeId },
-}
-
-#[derive(Serialize, Deserialize, Debug, Default, Clone, PartialEq, Eq)]
-pub struct JoinRequest {
-    pub node_id: NodeId,
-    pub endpoint: Endpoint,
-
-    #[serde(skip)]
-    #[deprecated(note = "it is listening addr, not advertise addr")]
-    pub grpc_api_addr: String,
-
-    pub grpc_api_advertise_address: Option<String>,
-}
-
-impl JoinRequest {
-    pub fn new(
-        node_id: NodeId,
-        endpoint: Endpoint,
-        grpc_api_advertise_address: Option<impl ToString>,
-    ) -> Self {
-        Self {
-            node_id,
-            endpoint,
-            grpc_api_advertise_address: grpc_api_advertise_address.map(|x| x.to_string()),
-            ..Default::default()
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct LeaveRequest {
-    pub node_id: NodeId,
-}
-
-#[derive(
-    Serialize, Deserialize, Debug, Clone, PartialEq, Eq, derive_more::From, derive_more::TryInto,
-)]
-pub enum ForwardRequestBody {
-    Ping,
-
-    Join(JoinRequest),
-    Leave(LeaveRequest),
-
-    Write(LogEntry),
-
-    GetKV(GetKVReq),
-    MGetKV(MGetKVReq),
-    ListKV(ListKVReq),
-}
-
-/// A request that is forwarded from one raft node to another
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct ForwardRequest {
-    /// Forward the request to leader if the node received this request is not leader.
-    pub forward_to_leader: u64,
-
-    pub body: ForwardRequestBody,
-}
-
-impl ForwardRequest {
-    pub fn decr_forward(&mut self) {
-        self.forward_to_leader -= 1;
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, derive_more::TryInto)]
-#[allow(clippy::large_enum_variant)]
-pub enum ForwardResponse {
-    #[try_into(ignore)]
-    Pong,
-
-    Join(()),
-    Leave(()),
-    AppliedState(AppliedState),
-
-    GetKV(GetKVReply),
-    MGetKV(MGetKVReply),
-    ListKV(ListKVReply),
-}
-
-impl tonic::IntoRequest<RaftRequest> for ForwardRequest {
-    fn into_request(self) -> tonic::Request<RaftRequest> {
-        let mes = RaftRequest {
-            data: serde_json::to_string(&self).expect("fail to serialize"),
-        };
-        tonic::Request::new(mes)
-    }
-}
-
-impl TryFrom<RaftRequest> for ForwardRequest {
-    type Error = tonic::Status;
-
-    fn try_from(mes: RaftRequest) -> Result<Self, Self::Error> {
-        let req = serde_json::from_str(&mes.data)
-            .map_err(|e| tonic::Status::invalid_argument(e.to_string()))?;
-        Ok(req)
-    }
-}
 
 impl tonic::IntoRequest<RaftRequest> for LogEntry {
     fn into_request(self) -> tonic::Request<RaftRequest> {
@@ -211,25 +95,6 @@ impl tonic::IntoRequest<RaftRequest> for &VoteRequest {
             data: serde_json::to_string(self).expect("fail to serialize"),
         };
         tonic::Request::new(mes)
-    }
-}
-impl From<RetryableError> for RaftReply {
-    fn from(err: RetryableError) -> Self {
-        let error = serde_json::to_string(&err).expect("fail to serialize");
-        RaftReply {
-            data: "".to_string(),
-            error,
-        }
-    }
-}
-
-impl From<AppliedState> for RaftReply {
-    fn from(msg: AppliedState) -> Self {
-        let data = serde_json::to_string(&msg).expect("fail to serialize");
-        RaftReply {
-            data,
-            error: "".to_string(),
-        }
     }
 }
